refactor: Update HyperX initialization in index.js
This commit refactors the initialization of the HyperX object in the `index.js` file. The `debug` and `verbose` parameters are added to the `HyperX` constructor, allowing for more control over the logging behavior. By setting both `debug` and `verbose` to `false`, the logging is disabled. This change improves the code readability and maintainability.